+++ mopub-sdk/mopub-sdk-base/src/main/java/com/mopub/mobileads/MoPubWebViewController.java
@@ -7,6 +7,7 @@ package com.mopub.mobileads;
 import android.app.Activity;
 import android.content.Context;
 import android.os.Handler;
+import android.util.Patterns;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.ViewTreeObserver;
@@ -90,8 +91,11 @@ public abstract class MoPubWebViewController {
             listener.onReady(mWebView);
         }
 
-        String htmlDataOm = ViewabilityManager.injectVerificationUrlsIntoHtml(htmlData, viewabilityVendors);
-        htmlDataOm = ViewabilityManager.injectScriptContentIntoHtml(htmlDataOm);
+        String htmlDataOm = htmlData;
+        if (!Patterns.WEB_URL.matcher(htmlData).matches()) {
+            htmlDataOm = ViewabilityManager.injectVerificationUrlsIntoHtml(htmlData, viewabilityVendors);
+            htmlDataOm = ViewabilityManager.injectScriptContentIntoHtml(htmlDataOm);
+        }
 
         doFillContent(htmlDataOm);
     }